cleaning up repo
