+++ source/index.html
@@ -8,9 +8,6 @@
     <meta name="author" content="Group 26">
     <meta name="description" content="CSE 110 Project">
     <link rel="stylesheet" href="style/menustyle.css">
-    <link rel="stylesheet" href="style/sidebarstyle.css">
-    <link rel="stylesheet" href="style/recipestyle.css">
-    <link rel="stylesheet" href="style/recipepage.style.css">
     <link rel="icon" type="image/png" href="/specs/researchmaterial/veganlogo.png">
 
     <meta charset="UTF-8" />
@@ -21,17 +18,12 @@
     <!-- Web Components -->
     <script src="scripts/RecipeCard.js" type="module"></script>
 
-    <!-- Service Worker -->
-    <!-- <script src="scripts/sw.js" type="module"></script> -->
-
-    <!-- Router -->
-    <script src="scripts/Router.js" type="module"></script>
-
     <!-- Main Stylesheet & Scripts -->
     <link id="main-theme" rel="stylesheet" href="style/menustyle.css" />
     <link id="theme" rel="stylesheet" type="text/css" href="" />
 
     <script src="scripts/main.js" type="module"></script>
+    <script src="scripts/search.js" type="module"></script>
 
 </head>
