--- source/recipes/krabby-patty.json
@@ -1,1 +1,0 @@
-{"@context":"http://schema.org","url":"https://www.delish.com/cooking/recipe-ideas/recipes/a48056/spongebob-inspired-krabby-patties-recipe/","publisher":{"@type":"Organization","name":"Delish","logo":{"@type":"ImageObject","url":"https://assets.hearstapps.com/sites/delish/assets/images/logos/logo-jsonld.58eed96.png","width":219,"height":60}},"@type":"Recipe","author":{"name":"Candace Braun Davison","@type":"Person"},"datePublished":"2016-08-11T03:08:00Z","headline":"Spongebob-Inspired Krabby Patties","image":{"@type":"ImageObject","height":1080,"thumbnail":"https://hips.hearstapps.com/del.h-cdn.co/assets/16/32/480x480/square-1470884407-delish-krabby-patty-02.jpg?resize=100:*","url":"https://hips.hearstapps.com/del.h-cdn.co/assets/16/32/1470884407-delish-krabby-patty-02.jpg","width":1920},"mainEntityOfPage":{"@id":"https://www.delish.com/cooking/recipe-ideas/recipes/a48056/spongebob-inspired-krabby-patties-recipe/","@type":"WebPage"},"thumbnailUrl":"https://hips.hearstapps.com/del.h-cdn.co/assets/16/32/480x480/square-1470884407-delish-krabby-patty-02.jpg?resize=100:*","isAccessibleForFree":"False","hasPart":[{"@type":"WebPageElement","isAccessibleForFree":"False","cssSelector":".content-container"}],"name":"Spongebob-Inspired Krabby Patties","prepTime":"PT0S","cookTime":"PT0S","totalTime":"PT20M","recipeIngredient":["1 1/3 lb. <p>ground beef (80 percent lean)</p>","2 tsp. <p>Old Bay seasoning</p>","1 tsp. <p>Black pepper</p>","1 tsp. <p>Sea salt</p>","4 <p>slices cheddar cheese</p>","1/2 c. <p>mayonnaise</p>","1 tsp. <p>Old Bay seasoning</p>","4 <p>sesame seed buns, sliced in half</p>","<p>ketchup</p>","<p>mustard</p>","1 <p>jar dill pickle slices</p>","1 <p>red onion, sliced</p>","1 <p>tomato, sliced</p>","4 <p>leaves of butter lettuce</p>"],"recipeInstructions":"<strong>Make the Burger Patties:</strong> Heat a grill pan or saut\u00e9 pan over medium heat. As it heats, combine ground beef, Old Bay, pepper and salt. Form into four equally sized patties, about 1/3 lb apiece. Set on a platter next to the stovetop. Coat the warmed grill pan with cooking spray and place patties on the pan, cooking each side about 4-5 minutes, or until cooked through to your desired doneness.  Use a large star-shaped cookie cutter to cut each cheese slice into a star shape. Place the star-shaped cheese on each burger patty, in the last 30 seconds or so of cooking, so it melts a little. Set burger patties aside. <strong>Make King Neptune\u2019s Poseidon Powder Aioli:</strong> Stir mayo and Old Bay until thoroughly combined.  <strong>Assemble the Sandwich:</strong> Spread Poseidon Powder Aioli on the bottom of each bun. Top with a burger patty, ketchup, mustard, pickles, onion slices, tomato and lettuce. Place the top bun on each burger. Serve.","video":{"@type":"VideoObject","contentUrl":"https://streaming.hearstdigitalstudios.com/manual_upload/57acfb4bd559300dad53407a/mp4/57acfb4bd559300dad53407a_Krabby_Patty_V4_1470954321_4000_hd_v3.mp4","description":"Plankton would kill to get his hands on this copycat krabby patty. The starfish-shaped cheese is an absolute must","duration":"PT59S","embedUrl":"https://glimmer.hearstapps.com/?id=4c65bae0-0e6a-4a95-a2d7-30c22497cad4","name":"Krabby Patty","thumbnailUrl":"https://hips.hearstapps.com/vidthumb/manual_upload/57acfb4bd559300dad53407a/thumb_1470954316.png","uploadDate":"2016-08-11T22:25:15.686873Z"},"recipeCuisine":[],"aggregateRating":{"@context":"http://schema.org/","@type":"AggregateRating","ratingValue":4.2222223,"reviewCount":9,"worstRating":1,"bestRating":5},"review":[{"@context":"http://schema.org/","@type":"Review","datePublished":"2020-07-15","author":{"@context":"http://schema.org/","@type":"Person","name":"Bubble Bass","givenName":"Bubble","familyName":"Bass"},"reviewBody":"The pickles go underneath the cheese because the cheese in the blanket on top of the pickles","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":2,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2020-07-19","author":{"@context":"http://schema.org/","@type":"Person","name":"Nope yeet","givenName":"Nope","familyName":"yeet"},"reviewBody":"Best krabby patty ever I love the recipe","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2020-10-27","author":{"@context":"http://schema.org/","@type":"Person","name":"max9362"},"reviewBody":"I loved the burger and this was my family when eating them, at least 40 dollars were spend on making them.","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2020-07-26","author":{"@context":"http://schema.org/","@type":"Person","name":"joe espocald","givenName":"joe","familyName":"espocald"},"reviewBody":"great tasting burger thanks for the recipe","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":1,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2019-01-12","author":{"@context":"http://schema.org/","@type":"Person","name":"GreenPopsicle"},"reviewBody":"The only adjustment I made was that I saut\u00e9ed the onions a bit, I'm not a fan of raw onion. They were absolutely fantastic! The Aioli was perfect, the seasoning gave the burger a wonderful savory twist the whole family (even my super finicky youngest son) loved! I'm definitely adding these to my picnic arsenal. Thanks!!","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2019-12-09","author":{"@context":"http://schema.org/","@type":"Person","name":"PlanktonsFriend"},"reviewBody":"I'm sending this to Plankton right now","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2019-11-24","author":{"@context":"http://schema.org/","@type":"Person","name":"GreenPalmtree"},"reviewBody":"this tastes so good that i vomited and drank my vomit over and over again, also i like it with extra sweat and toe nails thank yo for the recipe","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2019-11-24","author":{"@context":"http://schema.org/","@type":"Person","name":"FlopMaster44"},"reviewBody":"This was so good when i pooped it tasted like onions","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}},{"@context":"http://schema.org/","@type":"Review","datePublished":"2019-11-24","author":{"@context":"http://schema.org/","@type":"Person","name":"FlopMaster44"},"reviewBody":"5stars","reviewRating":{"@context":"http://schema.org/","@type":"Rating","ratingValue":5,"worstRating":1,"bestRating":5}}],"recipeCategory":["dinner","lunch","main dish"],"recipeYield":"4","description":"Take a visit to the Krusty Krab with this Spongebob-Inspired Krabby Patties recipe from Delish.com. ","keywords":"Food,Recipes"}
