--- source/recipes/noodle-soup.json
@@ -1,1 +1,0 @@
-{"@context":"http:\/\/schema.org","@type":"Recipe","name":"Kung Fu Panda Secret Ingredient Noodle Soup","author":{"@type":"Person","name":"Simply Inspired Meals"},"datePublished":"2021-01-16","recipeYield":4,"description":"Secret Ingredient Noodle Soup inspired by Kung Fu Panda&#x27;s Mr. Ping. This is an easy to make soup that tastes delicious, perfect for a cool evening.","image":["https:\/\/simplyinspiredmeals.com\/wp-content\/uploads\/2018\/09\/Kung-Fu-Panda-Secret-Ingredient-Noodle-Soup-320x320.jpg","https:\/\/simplyinspiredmeals.com\/wp-content\/uploads\/2018\/09\/Kung-Fu-Panda-Secret-Ingredient-Noodle-Soup-480x360.jpg","https:\/\/simplyinspiredmeals.com\/wp-content\/uploads\/2018\/09\/Kung-Fu-Panda-Secret-Ingredient-Noodle-Soup-480x270.jpg","https:\/\/simplyinspiredmeals.com\/wp-content\/uploads\/2018\/09\/Kung-Fu-Panda-Secret-Ingredient-Noodle-Soup.jpg"],"recipeCategory":"Soup","recipeCuisine":"Chinese","prepTime":"PT10M","cookTime":"PT15M","performTime":"PT15M","totalTime":"PT25M","recipeIngredient":["2 carrots, peeled and large diced","1\/4 of one onion, diced","3 stalks celery, large diced","1 clove garlic, minced","pinch of salt","pinch of black pepper","1 1\/2 tbsp. canola or vegetable oil","6 cups chicken broth","1 chicken bouillon cube","1\/2 cup water","1 tbsp soy sauce","2 packages chow mein noodles"],"recipeInstructions":[{"@type":"HowToStep","text":"In a large soup pot, combine the carrot, onion, green onion, celery, garlic and oil and cook until tender. Add the salt and pepper and mix well.","position":1,"name":"In a large soup pot, combine the carrot,...","url":"https:\/\/simplyinspiredmeals.com\/kung-fu-panda-secret-ingredient-noodle-soup\/#mv_create_54_1"},{"@type":"HowToStep","text":"Add broth, bouillon cube and water and bring to a boil.","position":2,"name":"Add broth, bouillon cube and water and bring...","url":"https:\/\/simplyinspiredmeals.com\/kung-fu-panda-secret-ingredient-noodle-soup\/#mv_create_54_2"},{"@type":"HowToStep","text":"Add soy sauce and noodles to the broth and cook for 5 to 10 minutes, stirring occasionally, until noodles are ready.","position":3,"name":"Add soy sauce and noodles to the broth...","url":"https:\/\/simplyinspiredmeals.com\/kung-fu-panda-secret-ingredient-noodle-soup\/#mv_create_54_3"},{"@type":"HowToStep","text":"Serve immediately.","position":4,"name":"Serve immediately.","url":"https:\/\/simplyinspiredmeals.com\/kung-fu-panda-secret-ingredient-noodle-soup\/#mv_create_54_4"}],"keywords":"#kungfupanda #kungfupandaparty #kungfupandafood #noodlesoup kung fu panda noodle soup, kung fu panda noodles, kung fu panda noodle soup recipe, kung fu panda soup, kung fu panda noodles recipe, kung fu panda noodle recipe, secret ingredient soup kung fu panda, kung fu panda secret ingredient soup recipe, noodle soup kung fu panda, mr. ping's noodle soup recipe, kung fu panda recipes, noodles from kung fu panda, kung fu panda food recipes","nutrition":{"@type":"NutritionInformation","calories":"228 calories","carbohydrateContent":"28 grams carbohydrates","cholesterolContent":"8 milligrams cholesterol","fatContent":"10 grams fat","fiberContent":"3 grams fiber","proteinContent":"6 grams protein","saturatedFatContent":"1 grams saturated fat","servingSize":"1","sodiumContent":"2126 grams sodium","sugarContent":"6 grams sugar","transFatContent":"0 grams trans fat","unsaturatedFatContent":"8 grams unsaturated fat"},"aggregateRating":{"@type":"AggregateRating","ratingValue":"4.7","reviewCount":12},"url":"https:\/\/simplyinspiredmeals.com\/kung-fu-panda-secret-ingredient-noodle-soup\/"}
