--- source/recipes/sweetrolls.json
@@ -1,1 +1,0 @@
-{"@context":"https://schema.org","@graph":[{"@type":"WebSite","@id":"https://sweettoothandsass.com/#website","url":"https://sweettoothandsass.com/","name":"Sweet Tooth and Sass","description":"A blog about all my baking endeavors!","publisher":{"@id":"https://sweettoothandsass.com/#/schema/person/4c2e703a0a856e2c07004de578e0aafe"},"potentialAction":[{"@type":"SearchAction","target":{"@type":"EntryPoint","urlTemplate":"https://sweettoothandsass.com/?s={search_term_string}"},"query-input":"required name=search_term_string"}],"inLanguage":"en-US"},{"@type":"ImageObject","@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#primaryimage","inLanguage":"en-US","url":"https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4.jpeg","contentUrl":"https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4.jpeg","width":481,"height":640,"caption":"Skyrim sweet rolls"},{"@type":"WebPage","@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#webpage","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/","name":"Skyrim Sweet Rolls - Sweet Tooth and Sass","isPartOf":{"@id":"https://sweettoothandsass.com/#website"},"primaryImageOfPage":{"@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#primaryimage"},"datePublished":"2021-05-07T22:20:58+00:00","dateModified":"2021-05-07T22:22:48+00:00","description":"Moist Skyrim sweet rolls drizzled with white icing. A lovely little dessert inspired by the The Elder scrolls video game.","breadcrumb":{"@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#breadcrumb"},"inLanguage":"en-US","potentialAction":[{"@type":"ReadAction","target":["https://sweettoothandsass.com/skyrim-sweet-rolls/"]}]},{"@type":"BreadcrumbList","@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#breadcrumb","itemListElement":[{"@type":"ListItem","position":1,"name":"Home","item":"https://sweettoothandsass.com/"},{"@type":"ListItem","position":2,"name":"Skyrim Sweet Rolls"}]},{"@type":"Article","@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#article","isPartOf":{"@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#webpage"},"author":{"@id":"https://sweettoothandsass.com/#/schema/person/4c2e703a0a856e2c07004de578e0aafe"},"headline":"Skyrim Sweet Rolls","datePublished":"2021-05-07T22:20:58+00:00","dateModified":"2021-05-07T22:22:48+00:00","wordCount":774,"commentCount":0,"publisher":{"@id":"https://sweettoothandsass.com/#/schema/person/4c2e703a0a856e2c07004de578e0aafe"},"image":{"@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#primaryimage"},"thumbnailUrl":"https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4.jpeg","articleSection":["Breakfast","Recipes"],"inLanguage":"en-US","potentialAction":[{"@type":"CommentAction","name":"Comment","target":["https://sweettoothandsass.com/skyrim-sweet-rolls/#respond"]}]},{"@type":["Person","Organization"],"@id":"https://sweettoothandsass.com/#/schema/person/4c2e703a0a856e2c07004de578e0aafe","name":"Carrie","image":{"@type":"ImageObject","@id":"https://sweettoothandsass.com/#personlogo","inLanguage":"en-US","url":"https://secure.gravatar.com/avatar/7d8bd31ab1e9fe933c99750f8554631c?s=96&d=mm&r=g","contentUrl":"https://secure.gravatar.com/avatar/7d8bd31ab1e9fe933c99750f8554631c?s=96&d=mm&r=g","caption":"Carrie"},"logo":{"@id":"https://sweettoothandsass.com/#personlogo"}},{"@context":"http://schema.org/","@type":"Recipe","name":"Skyrim sweet rolls","author":{"@type":"Person","name":"Carrie"},"description":"Moist sweet rolls drizzled in a white icing just like in the game!","datePublished":"2021-05-07T17:20:58+00:00","image":["https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4.jpeg","https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4-481x500.jpeg","https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4-481x375.jpeg","https://sweettoothandsass.com/wp-content/uploads/2021/05/ssw4-480x270.jpeg"],"recipeYield":["9","9 people"],"prepTime":"PT15M","cookTime":"PT20M","totalTime":"PT35M","recipeIngredient":["1 cup unsalted butter (room temperature)","1\u2154 cup granulated sugar","2 cups all-purpose flour","5 large eggs (room temperature)","\u00bd tsp salt","1 tsp baking powder","2 tsp vanilla extract","\u00bc cup heavy cream","4 tbsp unsalted butter","2 tbsp milk","2 cups confectioner&#039;s sugar","1 tsp vanilla extract"],"recipeInstructions":[{"@type":"HowToSection","name":"Skyrim sweet rolls","itemListElement":[{"@type":"HowToStep","text":"Preheat oven to 350\u00b0. Grease a mini bundt pan with non-stick spray and set aside.","name":"Preheat oven to 350\u00b0. Grease a mini bundt pan with non-stick spray and set aside.","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/#wprm-recipe-6245-step-0-0"},{"@type":"HowToStep","text":"In a bowl of a stand-mixer fitted with a paddle attachment, cream together the butter and sugar for 5 minutes until super light and fluffy.","name":"In a bowl of a stand-mixer fitted with a paddle attachment, cream together the butter and sugar for 5 minutes until super light and fluffy.","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/#wprm-recipe-6245-step-0-1"},{"@type":"HowToStep","text":"Add in one egg at a time, beating well after each addition. Scraping the sides of the bowl as needed. Add in the heavy cream and mix until blended.","name":"Add in one egg at a time, beating well after each addition. Scraping the sides of the bowl as needed. Add in the heavy cream and mix until blended.","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/#wprm-recipe-6245-step-0-2"},{"@type":"HowToStep","text":"Add in the flour, salt, baking powder, and vanilla extract and gently mix on low until just combined.","name":"Add in the flour, salt, baking powder, and vanilla extract and gently mix on low until just combined.","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/#wprm-recipe-6245-step-0-3"},{"@type":"HowToStep","text":"Pour batter into greased mini bundt pan and bake for 20 minutes or until golden brown around the bottom edges. Allow to cool for 5 minutes and then transfer to a cooking rack to complete cooling.","name":"Pour batter into greased mini bundt pan and bake for 20 minutes or until golden brown around the bottom edges. Allow to cool for 5 minutes and then transfer to a cooking rack to complete cooling.","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/#wprm-recipe-6245-step-0-4"}]},{"@type":"HowToSection","name":"White icing/glaze","itemListElement":[{"@type":"HowToStep","text":"Melt the butter in a microwavable bowl and add in the milk. Slowly, mix in the confectioner&#039;s sugar and vanilla. Drizzle over cooled sweet rolls and enjoy!","name":"Melt the butter in a microwavable bowl and add in the milk. Slowly, mix in the confectioner&#039;s sugar and vanilla. Drizzle over cooled sweet rolls and enjoy!","url":"https://sweettoothandsass.com/skyrim-sweet-rolls/#wprm-recipe-6245-step-1-0"}]}],"recipeCategory":["Breakfast","Dessert"],"recipeCuisine":["American"],"keywords":"Skyrim game, sweet rolls, breakfast","nutrition":{"@type":"NutritionInformation","servingSize":"9 people","calories":"642 kcal","carbohydrateContent":"86 g","proteinContent":"7 g","fatContent":"31 g","saturatedFatContent":"19 g","transFatContent":"1 g","cholesterolContent":"180 mg","sodiumContent":"225 mg","fiberContent":"1 g","sugarContent":"64 g"},"@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#recipe","isPartOf":{"@id":"https://sweettoothandsass.com/skyrim-sweet-rolls/#article"},"mainEntityOfPage":"https://sweettoothandsass.com/skyrim-sweet-rolls/#webpage"}]}
