--- source/recipes/thanos-krispies.json
@@ -1,1 +1,0 @@
-{"@context":"http:\/\/schema.org","@type":"Recipe","name":"Thanos Rice Krispie Infinity Stones","author":{"@type":"Person","name":"Kim and Kalee"},"datePublished":"2021-06-01","recipeYield":6,"description":"Any Marvel fan will love these Thanos Rice Krispie Treat Insanity Stones for an Avenger&#x27;s movie night! You can harness the power of the Infinity Gauntlet anytime you want with these colorful and delicious Rice Krispie treats!","image":["https:\/\/www.movienightsathome.com\/wp-content\/uploads\/2021\/06\/thanos-infinity-stones-rice-krispy-treats-min-720x720.jpg","https:\/\/www.movienightsathome.com\/wp-content\/uploads\/2021\/06\/thanos-infinity-stones-rice-krispy-treats-min-720x540.jpg","https:\/\/www.movienightsathome.com\/wp-content\/uploads\/2021\/06\/thanos-infinity-stones-rice-krispy-treats-min-720x405.jpg","https:\/\/www.movienightsathome.com\/wp-content\/uploads\/2021\/06\/thanos-infinity-stones-rice-krispy-treats-min.jpg"],"recipeCategory":"Movie Night Food Ideas","recipeCuisine":"Dessert","prepTime":"PT5M","cookTime":"PT15M","performTime":"PT15M","totalTime":"PT20M","recipeIngredient":["6 cups Kellogg\u2019s Rice Krispies Cereal","6 tablespoons Butter\u00a0","6 cups Mini Marshmallows","Food Coloring"],"recipeInstructions":[{"@type":"HowToStep","text":"Melt 1 tbsp. butter in a medium pan on medium to low heat","position":1,"name":"Melt 1 tbsp. butter in a medium pan...","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_1"},{"@type":"HowToStep","text":"Add 1 cup of miniature marshmallows and stir until the marshmallows are completely melted","position":2,"name":"Add 1 cup of miniature marshmallows and stir...","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_2"},{"@type":"HowToStep","text":"Take the pan off the heat.","position":3,"name":"Take the pan off the heat.","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_3"},{"@type":"HowToStep","text":"Add food coloring and mix until it's through the mixture.","position":4,"name":"Add food coloring and mix until it's through...","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_4"},{"@type":"HowToStep","text":"Add 1 cup of Rice Krispie Cereal and stir until it's mixed through.","position":5,"name":"Add 1 cup of Rice Krispie Cereal and...","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_5"},{"@type":"HowToStep","text":"Once the mixture is warm (not hot), make a stone out of each one with your hands.  You may want to spray your hands with non-stick spray first to help with the stickiness.","position":6,"name":"Once the mixture is warm (not hot), make...","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_6"},{"@type":"HowToStep","text":"Repeat for each color.","position":7,"name":"Repeat for each color.","url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/#mv_create_38_7"}],"keywords":"Thanos Rice Krispie Infinity Stones","nutrition":{"@type":"NutritionInformation","calories":"366 calories","carbohydrateContent":"64 grams carbohydrates","cholesterolContent":"31 milligrams cholesterol","fatContent":"12 grams fat","fiberContent":"0 grams fiber","proteinContent":"3 grams protein","saturatedFatContent":"7 grams saturated fat","servingSize":"1","sodiumContent":"277 grams sodium","sugarContent":"32 grams sugar","transFatContent":"0 grams trans fat","unsaturatedFatContent":"4 grams unsaturated fat"},"aggregateRating":{"@type":"AggregateRating","ratingValue":5,"reviewCount":"1"},"url":"https:\/\/www.movienightsathome.com\/thanos-rice-krispie-infinity-stones\/"}
